saved everything i could
I had to go onto the github website and download different clones of our
project at different times to find the last save that had a working game
scene and copy it into our project with the latest scripts. I'm not sure
what happened but something happened to the scene file on Jerry's last
commit. Github must have just been being a bitch. I'm not sure what
Jerry is doing with the inventory but it is no longer working.