NewInventorySystem
i've implemented some functionality to the alternative inventory system
we have talked about in class. You can access it through the supply
storage room. When you send out scavengers they will generate a random
stack of a random item. When you click on an item it just decrements the
amount of that item you are carrying.