updated Info panel!!!
i've added a short description to each item and the name to pop up as
well. I've also set some ground work for the crafting system, you can
select the use for crafting button to send it to one of the crafting
nodes in my TestingInventory gameObject to the node variables I have set
up there.