Bug Fixes and Inventory Help Panels
I've created two new help panels for inventory and crafting. Inventory
and crafting are now assigned to their respective rooms but the supply
storage is now pointless. I also fixed some bugs that showed up like not
spending a molotov when it says you did and being able to select a room
when the notification window is open.